Using Celery signal for race condition.
Looks like the Celery worker in the background thread was still
initializing SQL tables when tests began to run on AppVeyor, since VMs
on that CI are very slow.

Using the Celery signal to fire off a function once the thread is done
initializing, giving tests up to 20 seconds before they start running.
Seems to have fixed the issue.